Fix: Home 페이지 설정
